Implement support for extracting diagonal from kernel
This is requested through to AST optimisation option extract_diagonal
and is only meaningful for kernel that produce a square local matrix.

If enabled, the double loop nest inside the kernel is replaced by a
single loop and indices using the iteration variable of the inner
loop are replace by the iteration variable of the outer loop. The
output argument is replaced by a vector of the side length of the
local matrix of the orginal kernel.